Добавляет модуль с картой. Описывает поведения маркеров и появление балунов. Много комментов в коде
