+++ js/map.js
@@ -1,80 +1,151 @@
 /* global L:readonly */
-import { createAdverts } from './data.js';
+// import { createAdverts } from './data.js';
 import { createCard } from './adverts.js';
-const similarAdverts = createAdverts();
+// const similarAdverts = createAdverts();
 
 
-const adForm = document.querySelector('.ad-form');
-const adFormFieldsets = document.querySelectorAll('.ad-form fieldset');
-const filtersForm = document.querySelector('.map__filters');
-const filtersFormSelects = document.querySelectorAll('.map__filters select');
-const filtersFormFieldset = document.querySelector('#housing-features');
-const CITY_CENTER = [35.686427, 139.753637];
+// const adForm = document.querySelector('.ad-form');
+// const adFormFieldsets = document.querySelectorAll('.ad-form fieldset');
+// const filtersForm = document.querySelector('.map__filters');
+// const filtersFormSelects = document.querySelectorAll('.map__filters select');
+// const filtersFormFieldset = document.querySelector('#housing-features');
+// const CITY_CENTER = [35.686427, 139.753637];
 
 /**Деактивирует форму */
-adForm.classList.add('ad-form--disabled');
+const disactivateForm = (adForm, adFormFieldsets, filtersForm, filtersFormSelects, filtersFormFieldset) => {
+  adForm.classList.add('ad-form--disabled');
 
-for (let fieldset of adFormFieldsets) {
-  fieldset.setAttribute('disabled', 'disabled');
+  for (let fieldset of adFormFieldsets) {
+    fieldset.setAttribute('disabled', 'disabled');
+  }
+
+  filtersForm.classList.add('map__filters--disabled');
+
+  for (let select of filtersFormSelects) {
+    select.setAttribute('disabled', 'disabled');
+  }
+
+  filtersFormFieldset.setAttribute('disabled', 'disabled');
+};
+
+/**Активирует форму */
+const activateForm = (adForm, adFormFieldsets, filtersForm, filtersFormSelects, filtersFormFieldset) => {
+  adForm.classList.remove('ad-form--disabled');
+
+  for (let fieldset of adFormFieldsets) {
+    fieldset.removeAttribute('disabled', 'disabled');
+  }
+
+  filtersForm.classList.remove('map__filters--disabled');
+
+  for (let select of filtersFormSelects) {
+    select.removeAttribute('disabled', 'disabled');
+  }
+
+  filtersFormFieldset.removeAttribute('disabled', 'disabled');
 }
 
-filtersForm.classList.add('map__filters--disabled');
+/**
+ * Добавляет карту и активирует форму
+ * @param onLoadFunc Функция при загрузке карты.
+ * @param coordinates Массив из 2х координат lat & lng
+*/
+const initMap = (onLoadFunc, coordinates) => {
+  const map = L.map('map-canvas')
+    .on('load', () => {
+      onLoadFunc;
+    })
 
-for (let select of filtersFormSelects) {
-  select.setAttribute('disabled', 'disabled');
-}
+    .setView({
+      lat: coordinates[0],
+      lng: coordinates[1],
+    }, 12);
 
-filtersFormFieldset.setAttribute('disabled', 'disabled');
+  L.tileLayer(
+    'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+    {
+      attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+    },
+  ).addTo(map);
+  return map;
+};
 
-/** Добавляет карту и активирует форму */
-const map = L.map('map-canvas')
-  .on('load', () => {
+// const map = L.map('map-canvas')
+//     .on('load', () => {
 
-    adForm.classList.remove('ad-form--disabled');
+//       adForm.classList.remove('ad-form--disabled');
 
-    for (let fieldset of adFormFieldsets) {
-      fieldset.removeAttribute('disabled', 'disabled');
-    }
+//       for (let fieldset of adFormFieldsets) {
+//         fieldset.removeAttribute('disabled', 'disabled');
+//       }
 
-    filtersForm.classList.remove('map__filters--disabled');
+//       filtersForm.classList.remove('map__filters--disabled');
 
-    for (let select of filtersFormSelects) {
-      select.removeAttribute('disabled', 'disabled');
-    }
+//       for (let select of filtersFormSelects) {
+//         select.removeAttribute('disabled', 'disabled');
+//       }
 
-    filtersFormFieldset.removeAttribute('disabled', 'disabled');
-  })
-  .setView({
-    lat: CITY_CENTER[0],
-    lng: CITY_CENTER[1],
-  }, 12);
+//       filtersFormFieldset.removeAttribute('disabled', 'disabled');
+//     })
 
-L.tileLayer(
-  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
-  {
-    attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
-  },
-).addTo(map);
+//     .setView({
+//       lat: CITY_CENTER[0],
+//       lng: CITY_CENTER[1],
+//     }, 12);
 
-//Добавляет главный маркер на карту
-const mainMarkerIcon = L.icon({
-  iconUrl: '../img/main-pin.svg',
-  iconSize: [52, 52],
-  iconAnchor: [26, 52],
-});
+//   L.tileLayer(
+//     'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+//     {
+//       attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+//     },
+//   ).addTo(map);
 
-const mainMarker = L.marker(
-  {
-    lat: CITY_CENTER[0],
-    lng: CITY_CENTER[1],
-  },
-  {
-    draggable: true,
-    icon: mainMarkerIcon,
-  },
-);
 
-mainMarker.addTo(map);
+
+/**
+ * Добавляет главный маркер на карту
+ * @param coordinates Массив из 2х координат lat & lng
+*/
+const addMainMarker = (coordinates, map) => {
+  const mainMarkerIcon = L.icon({
+    iconUrl: '../img/main-pin.svg',
+    iconSize: [52, 52],
+    iconAnchor: [26, 52],
+  });
+
+  const mainMarker = L.marker(
+    {
+      lat: coordinates[0],
+      lng: coordinates[1],
+    },
+    {
+      draggable: true,
+      icon: mainMarkerIcon,
+    },
+  );
+
+  mainMarker.addTo(map);
+  return mainMarker;
+};
+
+// const mainMarkerIcon = L.icon({
+//   iconUrl: '../img/main-pin.svg',
+//   iconSize: [52, 52],
+//   iconAnchor: [26, 52],
+// });
+
+// const mainMarker = L.marker(
+//   {
+//     lat: CITY_CENTER[0],
+//     lng: CITY_CENTER[1],
+//   },
+//   {
+//     draggable: true,
+//     icon: mainMarkerIcon,
+//   },
+// );
+
+// mainMarker.addTo(map);
 
 /**Получаем координаты маркера */
 
@@ -82,40 +153,81 @@ mainMarker.addTo(map);
 //   console.log(evt.target.getLatLng());
 // });
 
-/**Добавляем координаты остановки маркера в поле адрес */
+/**
+ * Добавляем координаты остановки маркера в поле адрес
+ * @param addressField Поле адреса из формы
+ * @param coordinates Начальные координаты маркера
+ * @param mainMarker Целевой маркер
+ */
+const catchMainMarkerCoordinates = (addressField, coordinates, mainMarker) => {
+  addressField.value = `${coordinates[0].toFixed(5)}, ${coordinates[1].toFixed(5)}`;
 
-const adFormAddress = document.querySelector('#address');
+  mainMarker.on('moveend', (evt) => {
+    let currentX = evt.target.getLatLng().lat.toFixed(5);
+    let currentY = evt.target.getLatLng().lng.toFixed(5);
+    addressField.value = `${currentX}, ${currentY}`;
+  });
+}
 
-adFormAddress.value = `${CITY_CENTER[0].toFixed(5)}, ${CITY_CENTER[1].toFixed(5)}`;
+// adFormAddress.value = `${CITY_CENTER[0].toFixed(5)}, ${CITY_CENTER[1].toFixed(5)}`;
 
-mainMarker.on('moveend', (evt) => {
-  let currentX = evt.target.getLatLng().lat.toFixed(5);
-  let currentY = evt.target.getLatLng().lng.toFixed(5);
-  adFormAddress.value = `${currentX}, ${currentY}`;
-});
+// mainMarker.on('moveend', (evt) => {
+//   let currentX = evt.target.getLatLng().lat.toFixed(5);
+//   let currentY = evt.target.getLatLng().lng.toFixed(5);
+//   adFormAddress.value = `${currentX}, ${currentY}`;
+// });
 
 
 /**Добавляет похожие маркеры на карту */
-similarAdverts.forEach((advert) => {
-  const markerIcon = L.icon({
-    iconUrl: '../img/pin.svg',
-    iconSize: [40, 40],
-    iconAnchor: [20, 40],
-  });
+const addSimilarMarkers = (similarAdverts, map) => {
+  similarAdverts.forEach((advert) => {
+    const markerIcon = L.icon({
+      iconUrl: '../img/pin.svg',
+      iconSize: [40, 40],
+      iconAnchor: [20, 40],
+    });
 
-  const marker = L.marker(
-    {
-      lat: advert.location.lat,
-      lng: advert.location.lng,
-    },
-    {
-      icon: markerIcon,
-    },
-  );
-
-  marker
-    .addTo(map)
-    .bindPopup(
-      createCard(advert),
+    const marker = L.marker(
+      {
+        lat: advert.location.lat,
+        lng: advert.location.lng,
+      },
+      {
+        icon: markerIcon,
+      },
     );
-});
+
+    marker
+      .addTo(map)
+      .bindPopup(
+        createCard(advert),
+      );
+  });
+};
+
+
+// similarAdverts.forEach((advert) => {
+//   const markerIcon = L.icon({
+//     iconUrl: '../img/pin.svg',
+//     iconSize: [40, 40],
+//     iconAnchor: [20, 40],
+//   });
+
+//   const marker = L.marker(
+//     {
+//       lat: advert.location.lat,
+//       lng: advert.location.lng,
+//     },
+//     {
+//       icon: markerIcon,
+//     },
+//   );
+
+//   marker
+//     .addTo(map)
+//     .bindPopup(
+//       createCard(advert),
+//     );
+// });
+
+export {disactivateForm, activateForm, initMap, addMainMarker, catchMainMarkerCoordinates, addSimilarMarkers};
